requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.7
dj-database-url==2.2.0
Django==4.2.13
djangorestframework==3.14.0
gunicorn==22.0.0
h11==0.14.0
packaging==24.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.2
uvicorn==0.30.1
whitenoise==6.6.0
